backend/lambda/rag-agent/requirements.txt
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.20
langchain-openfga>=0.1.0
openfga-sdk>=0.3.0
faiss-cpu>=1.7.4
boto3>=1.34.0
pyjwt[crypto]>=2.8.0
requests>=2.31.0
cachetools>=5.3.0
